% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{autoTransformFit}
\alias{autoTransformFit}
\title{Fit an autotransformed linear model}
\usage{
autoTransformFit(fit, vars2transform = NULL)
}
\arguments{
\item{fit}{An object of class lm}

\item{vars2transform}{Character vector Variable names to transform or NULL}
}
\description{
Fit an autotransformed linear model
}
\examples{
data(acs,package="moonBook")
fit<- lm(EF ~ weight+TG,data=acs)
fit<- lm(EF ~ age+TG,data=acs)  ## age donot need transformation
fit<- lm(age ~ EF+TG,data=acs)  ## age donot need transformation
fit<- lm(EF ~ age+TG+sex+DM,data=acs)  ## age, sex donot need transformation
result=autoTransformFit(fit)
fit<- lm(mpg ~ hp+wt,data=mtcars)
result=autoTransformFit(fit)
plot(result,fill="red")
data("autotrader",package="bestNormalize")
autotrader$yearsold <- 2017 - autotrader$Year
fit<- lm(price ~ mileage + yearsold+status, data = autotrader)
result=autoTransformFit(fit)
}
