% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassify.R
\name{boxPlot2}
\alias{boxPlot2}
\title{Draw a box plot}
\usage{
boxPlot2(y, pred, labels, ...)
}
\arguments{
\item{y}{A numeric vector}

\item{pred}{Numeric vecor as a result of predict.glm}

\item{...}{further arguments to be passed to geom_boxplot()}
}
\value{
A ggplot
}
\description{
Draw a box plot
}
\examples{
form2=paste0("AMD~",paste0(colnames(ExampleData)[3:16],collapse="+"))
fit2=glm(as.formula(form2),data=ExampleData,family=binomial)
pred2=predict(fit2,type="response")
boxPlot2(y=ExampleData$AMD,pred=pred2)
}
