% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassify.R
\name{cutoff2str}
\alias{cutoff2str}
\title{Convert cutoff to string}
\usage{
cutoff2str(cutoff = c(0, 0.1, 0.35, 1))
}
\arguments{
\item{cutoff}{Numeric vector}
}
\description{
Convert cutoff to string
}
\examples{
cutoff2str(cutoff=c(0,.1,.2,.5,1))
}
