% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassify.R
\name{ggplotROC}
\alias{ggplotROC}
\title{Draw ROC curves}
\usage{
ggplotROC(fit1, fit2, labels, ...)
}
\arguments{
\item{fit1, fit2}{An object of class glm}

\item{labels}{Character}

\item{...}{further arguments to be passed to geom_line()}
}
\description{
Draw ROC curves
}
\examples{
form1=paste0("AMD~",paste0(colnames(ExampleData)[3:10],collapse="+"))
form2=paste0("AMD~",paste0(colnames(ExampleData)[3:16],collapse="+"))
fit1=glm(as.formula(form1),data=ExampleData,family=binomial)
fit2=glm(as.formula(form2),data=ExampleData,family=binomial)
labels <- c("without genetic factors", "with genetic factors")
ggplotROC(fit1,fit2,labels=labels)
}
