% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictModel.R
\name{makeNewData2}
\alias{makeNewData2}
\title{Make new data for predict}
\usage{
makeNewData2(model, length = 100, by = NULL, type = "response")
}
\arguments{
\item{model}{An object}

\item{length}{numeric length of continuous variable to to predict}

\item{by}{character optional factor variable}

\item{type}{character type argument to be passed to predict.gam}
}
\description{
Make new data for predict
}
\examples{
model=lm(mpg~wt+hp+am,data=mtcars)
makeNewData2(model)
}
