% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTLIst_autoTransform.R
\name{makePPTList_autoTransform}
\alias{makePPTList_autoTransform}
\title{Make Powerpoint list for autoTransformFit}
\usage{
makePPTList_autoTransform(
  fit,
  vars2transform = NULL,
  vanilla = TRUE,
  show.transformDensity = TRUE,
  add.gam = TRUE
)
}
\arguments{
\item{fit}{An object of class lm}

\item{vars2transform}{Optional names of variables to trnasform}

\item{vanilla}{logical If true, make a vanilla table}

\item{show.transformDensity, add.gam}{logical Whether or not add tansformDensity or gam result}
}
\description{
Make Powerpoint list for autoTransformFit
}
\examples{
data(autotrader)
fit<- lm(price ~ mileage + yearsold +status, data= autotrader )
result=makePPTList_autoTransform(fit)
result=makePPTList_autoTransform(fit,vars2transform=c('mileage'))
result=makePPTList_autoTransform(fit,show.transformDensity=FALSE)
result=makePPTList_autoTransform(fit,add.gam=FALSE)
result=makePPTList_autoTransform(fit,show.transformDensity=FALSE,add.gam=FALSE)
}
