% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassify.R
\name{makePPTList_reclassfy}
\alias{makePPTList_reclassfy}
\title{Make Powerpoint list for reclassification}
\usage{
makePPTList_reclassfy(fit1, fit2, labels, cutoff = c(0, 0.1, 0.35, 1), cutoff2)
}
\arguments{
\item{fit1, fit2}{An object of class glm}

\item{labels}{Optioanl labels for initial model and updated model}

\item{cutoff, cutoff2}{A numeric vector Cutoff values for risk categories for initial and updated model}
}
\description{
Make Powerpoint list for reclassification
}
\examples{
form1=paste0("AMD~",paste0(colnames(ExampleData)[3:10],collapse="+"))
form2=paste0("AMD~",paste0(colnames(ExampleData)[3:16],collapse="+"))
fit1=glm(as.formula(form1),data=ExampleData,family=binomial)
fit2=glm(as.formula(form2),data=ExampleData,family=binomial)
labels=c("Without Genetic Factor","With Genetic Factor")
data=makePPTList_reclassfy(fit1,fit2,labels=labels)
}
