% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{needTransform}
\alias{needTransform}
\title{Check for need for transformation}
\usage{
needTransform(x, family = "bcnPower")
}
\arguments{
\item{x}{A numeric vector}

\item{family}{Character The quoted name of a family of transformations.}
}
\value{
A logical value
}
\description{
Check for need for transformation
}
\examples{
needTransform(mtcars$mpg)
needTransform(iris$Sepal.Length)
}
