% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassify.R
\name{ORmultivariate}
\alias{ORmultivariate}
\title{Function to obtain multivariate odds ratios from a logistic regression model.}
\usage{
ORmultivariate(fit)
}
\arguments{
\item{fit}{An object of class glm}
}
\value{
A list with elements
\describe{
\item{Predictors_Summary}{OR with 95\% CI and corresponding p-values for each predictor in the model}
\item{Brier_Score}{Brier score}
\item{Nagelkerke_Index}{Nagelkerke's R2 value}
}
}
\description{
Function to obtain multivariate odds ratios from a logistic regression model.
}
\examples{
form1=paste0("AMD~",paste0(colnames(ExampleData)[3:10],collapse="+"))
form2=paste0("AMD~",paste0(colnames(ExampleData)[3:16],collapse="+"))
fit1=glm(as.formula(form1),data=ExampleData,family=binomial)
fit2=glm(as.formula(form2),data=ExampleData,family=binomial)
ORmultivariate(fit1)
ORmultivariate(fit2)
}
\references{
Suman Kundu, Yurii S. Aulchenko and A. Cecile J.W. Janssens (2020). PredictABEL: Assessment of Risk Prediction Models. R package version 1.2-4.https://CRAN.R-project.org/package=PredictABEL
Brier GW. Verification of forecasts expressed in terms of probability. Monthly weather review 1950;78:1-3.
Nagelkerke NJ. A note on a general definition of the coefficient of determination. Biometrika 1991;78:691-692.
}
