% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{plot.autoBN}
\alias{plot.autoBN}
\title{S3 method of class autoBN}
\usage{
\method{plot}{autoBN}(x, ...)
}
\arguments{
\item{x}{An object of calss autoBN}

\item{...}{Further arguments to be passed to plot_autoBN}
}
\description{
S3 method of class autoBN
}
\examples{
data(acs,package="moonBook")
fit<- lm(EF ~ age+TG+sex,data=acs)  ## age, sex donot need transformation
result=autoTransformFit(fit)
plot(result,fill="red")
plot(result,select=1:2,fill="blue")
plot(result,add.gam=TRUE)
plot(result,by=sex)
plot(result,add.gam=TRUE,by=sex,select=1)
data("autotrader",package="bestNormalize")
autotrader$yearsold <- 2017 - autotrader$Year
fit<- lm(price ~ mileage + yearsold+status, data = autotrader)
result=autoTransformFit(fit)
plot(result,fill="red")
plot(result,add.gam=TRUE)
plot(result,add.gam=TRUE,select=1)
}
