% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{plot.BNdf}
\alias{plot.BNdf}
\title{S3 method of class BNdf}
\usage{
\method{plot}{BNdf}(x, ...)
}
\arguments{
\item{x}{An object of class BNdf}

\item{...}{Further arguments to be passed to plot_BNdf}
}
\description{
S3 method of class BNdf
}
\examples{
x=autoTransformDf(iris)
plot(x)
}
