% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{plot_autoBN}
\alias{plot_autoBN}
\title{Plot an object of class autoBN}
\usage{
plot_autoBN(x, add.gam, show.point, by, select, ...)
}
\arguments{
\item{x}{An object of calss autoBN}

\item{add.gam}{logical Whether or not add gam fit}

\item{show.point}{logical Whether or not show point}

\item{by}{Optional name of categorical variable}

\item{select}{Numeric Plot choices}

\item{...}{Further arguments to be passed to geom_ribbon}
}
\description{
Plot an object of class autoBN
}
