% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{plot_BNdf}
\alias{plot_BNdf}
\title{Plot function for an object of class BNdf}
\usage{
plot_BNdf(x, select = NULL)
}
\arguments{
\item{x}{An object of class BNdf}

\item{select}{Choices of BN to plot}
}
\description{
Plot function for an object of class BNdf
}
\examples{
x=autoTransformDf(iris)
plot_BNdf(x)
}
