% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{plot_BNdf2}
\alias{plot_BNdf2}
\title{BoxPlot function for an object of class BNdf}
\usage{
plot_BNdf2(x, select = NULL)
}
\arguments{
\item{x}{An object of class BNdf}

\item{select}{Choices of BN to plot}
}
\description{
BoxPlot function for an object of class BNdf
}
\examples{
x=autoTransformDf(iris)
plot_BNdf2(x)
}
