% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{plot_BNdf4}
\alias{plot_BNdf4}
\title{Predict plot function for an object of class BNdf}
\usage{
plot_BNdf4(x, select = NULL)
}
\arguments{
\item{x}{An object of class BNdf}

\item{select}{Choices of BN to plot}
}
\description{
Predict plot function for an object of class BNdf
}
\examples{
x=autoTransformDf(iris)
plot_BNdf4(x)
}
