% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictModel.R
\name{predictModel}
\alias{predictModel}
\title{Draw a ggplot with an object of class gam}
\usage{
predictModel(
  model,
  select = NULL,
  point = TRUE,
  se = TRUE,
  by,
  type = "response",
  interactive = FALSE,
  fillcolor = "red",
  fillalpha = 0.4,
  ...
)
}
\arguments{
\item{model}{An object of class gam}

\item{select}{numeric Choices of dependent variables to plot}

\item{point}{logical Whether or not draw point}

\item{se}{logical Whether or not draw confidence interval}

\item{by}{NULL or character optional name of factor variable}

\item{type}{character type argument to be passed to predict.gam}

\item{interactive}{logical If true, make a interactive plot}

\item{fillcolor}{character The fill color of geom_ribbon}

\item{fillalpha}{numeric The alpha value of geom_ribbon}

\item{...}{further arguments to be passed to geom_point_interactive}
}
\description{
Draw a ggplot with an object of class gam
}
\examples{
model=lm(mpg~wt+hp+am,data=mtcars)
predictModel(model,interactive=TRUE)
predictModel(model,by=am)
predictModel(model,by=am,interactive=TRUE)
\donttest{
model=lm(Sepal.Length~Sepal.Width+Petal.Length+Species,data=iris)
summary(model)
predictModel(model)
predictModel(model,interactive=TRUE)
predictModel(model,by=Species)
data(colon,package="survival")
model=glm(status~rx+age+nodes,data=colon,family=binomial)
predictModel(model)
predictModel(model,by=rx)
predictModel(model,by=rx,type="link")
predictModel(model,type="link")
}
}
