% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassify.R
\name{predictPlot}
\alias{predictPlot}
\title{Draw a Predictiveness Plot}
\usage{
predictPlot(predlist, labels, size, ...)
}
\arguments{
\item{predlist}{A list of prediction vectors}

\item{labels}{String Optional labels}

\item{...}{Further arguments to be passed to geom_line()}
}
\value{
A ggplot
}
\description{
Draw a Predictiveness Plot
}
\examples{
form1=paste0("AMD~",paste0(colnames(ExampleData)[3:10],collapse="+"))
form2=paste0("AMD~",paste0(colnames(ExampleData)[3:16],collapse="+"))
fit1=glm(as.formula(form1),data=ExampleData,family=binomial)
fit2=glm(as.formula(form2),data=ExampleData,family=binomial)
pred1=predict(fit1,type="response")
pred2=predict(fit2,type="response")
labels=c("without genetic factors", "with genetic factors")
predictPlot(list(pred1,pred2),labels=labels)
}
