% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassify.R
\name{print.reclassified}
\alias{print.reclassified}
\title{S3 method print for an object of class reclassified}
\usage{
\method{print}{reclassified}(x, ...)
}
\arguments{
\item{x}{An object of class reclassified}

\item{...}{Fruther arguments}
}
\description{
S3 method print for an object of class reclassified
}
