% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassify.R
\name{reclassify}
\alias{reclassify}
\title{Function for reclassification table and statistics}
\usage{
reclassify(fit1, fit2, cutoff, cutoff2)
}
\arguments{
\item{fit1, fit2}{An object of class glm}

\item{cutoff}{Cutoff values for risk categories for initial model. Default value is c(0,.1,.35,1)}

\item{cutoff2}{Optional cutoff values for risk categories for updated model}
}
\value{
A list of class reclassified with elements
\describe{
\item{cutoff}{Cutoff values for risk categories for initial model}
\item{cutoff2}{Cutoff values for risk categories for updated model}
\item{absent}{summary table for absent outcome}
\item{present}{summary table for present outcome}
\item{combined}{summary table for combined outcome}
\item{x}{improveProb for categorical variables}
\item{y}{improveProb for initial and updated model}
\item{NRIcat}{Numeric vector for Net reclassification improvement, categorical}
\item{NRIcont}{Numeric vector for Net reclassification improvement, continuous}
\item{IDI}{Numeric vector for Integrated discrimination improvement}
}
}
\description{
Function for reclassification table and statistics
}
\examples{
form1=paste0("AMD~",paste0(colnames(ExampleData)[3:10],collapse="+"))
form2=paste0("AMD~",paste0(colnames(ExampleData)[3:16],collapse="+"))
fit1=glm(as.formula(form1),data=ExampleData,family=binomial)
fit2=glm(as.formula(form2),data=ExampleData,family=binomial)
result=reclassify(fit1,fit2)
result
}
