% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassify.R
\name{reclassTable}
\alias{reclassTable}
\title{Make recalssification table}
\usage{
reclassTable(x, outcome, labels)
}
\arguments{
\item{x}{An object of class reclassified}

\item{outcome}{Character}

\item{labels}{Optional character}
}
\value{
An object of class flextable
}
\description{
Make recalssification table
}
\examples{
form1=paste0("AMD~",paste0(colnames(ExampleData)[3:10],collapse="+"))
form2=paste0("AMD~",paste0(colnames(ExampleData)[3:16],collapse="+"))
fit1=glm(as.formula(form1),data=ExampleData,family=binomial)
fit2=glm(as.formula(form2),data=ExampleData,family=binomial)
result=reclassify(fit1,fit2)
reclassTable(result)
}
