% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassify.R
\name{riskDistributionPlot}
\alias{riskDistributionPlot}
\title{Risk Distribution Plot}
\usage{
riskDistributionPlot(y, pred, labels, ...)
}
\arguments{
\item{y}{A numeric vector}

\item{pred}{Numeric vecor as a result of predict.glm}

\item{labels}{Character}

\item{...}{further arguments to be passed to geom_col()}
}
\value{
A ggplot
}
\description{
Risk Distribution Plot
}
\examples{
form2=paste0("AMD~",paste0(colnames(ExampleData)[3:16],collapse="+"))
fit2=glm(as.formula(form2),data=ExampleData,family=binomial)
pred2=predict(fit2,type="response")
riskDistributionPlot(y=ExampleData$AMD,pred=pred2)
}
