% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showStat.R
\name{showStat}
\alias{showStat}
\title{Draw distribution of statistics}
\usage{
showStat(
  fun = NULL,
  df1 = NULL,
  df2 = NULL,
  stat = NULL,
  alternative = c("two.sided", "less", "greater"),
  alpha = 0.05,
  show.area = TRUE,
  show.critical = TRUE,
  show.point = TRUE,
  color = "black",
  fill = "red",
  digits = 3
)
}
\arguments{
\item{fun}{distrubution function}

\item{df1, df2}{degrees of freedom}

\item{stat}{numeric The value of statistics}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}

\item{alpha}{numeric confidence level of the interval}

\item{show.area, show.critical, show.point}{logical}

\item{color}{character Name of line color}

\item{fill}{character Name of fill color}

\item{digits}{integer indicating the number of decimal places}
}
\description{
Draw distribution of statistics
}
\examples{
chisq.test(table(mtcars$am,mtcars$vs))
showStat(fun=dchisq,df1=1,stat=0.34754)
t.test(mpg~am,data=mtcars)
showStat(dt,df1=18.332,stat=-3.7671)
}
