% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformDensity.R
\name{transformDensity}
\alias{transformDensity}
\title{Draw histogram of transformed value with normal density}
\usage{
transformDensity(
  x,
  best = FALSE,
  power = TRUE,
  hist = TRUE,
  density = FALSE,
  select = NULL,
  method = c("shapiro", "ad", "cvm", "lillie", "pearson", "sf")
)
}
\arguments{
\item{x}{A numeric vector}

\item{best}{logical Whether or not use bestNormalize::bestNormalize}

\item{power}{logical Whether or not use car::powerTransform}

\item{hist}{logical Whether or not draw histogram}

\item{density}{logical Whether or not draw density}

\item{select}{numeric choice of normality test}

\item{method}{character choice of normality test}
}
\description{
Draw histogram of transformed value with normal density
}
\examples{
require(moonBook)
transformDensity(acs$TG)
transformDensity(acs$TG,best=TRUE)
}
