% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{autoTransformDf}
\alias{autoTransformDf}
\title{Autologous transformation of data.frame using bestNormalize}
\usage{
autoTransformDf(df)
}
\arguments{
\item{df}{A data.frame}
}
\description{
Autologous transformation of data.frame using bestNormalize
}
