% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassify.R
\name{calibrationPlot}
\alias{calibrationPlot}
\title{Draw a calibration plot}
\usage{
calibrationPlot(fit, ...)
}
\arguments{
\item{fit}{An object of class glm}

\item{...}{further arguments to be passed to geom_point()}
}
\value{
A ggplot
}
\description{
Draw a calibration plot
}
\examples{
form2=paste0("AMD~",paste0(colnames(ExampleData)[3:16],collapse="+"))
fit2=glm(as.formula(form2),data=ExampleData,family=binomial)
calibrationPlot(fit2)
}
