% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassify.R
\name{fit2call}
\alias{fit2call}
\title{Get call as string from an object of class glm or lm}
\usage{
fit2call(fit)
}
\arguments{
\item{fit}{an obkect of class glm}
}
\description{
Get call as string from an object of class glm or lm
}
\examples{
form1=paste0("AMD~",paste0(colnames(ExampleData)[3:10],collapse="+"))
fit=glm(as.formula(form1),data=ExampleData,family=binomial)
fit2call(fit)
fit<- lm(mpg~hp*wt,data=mtcars)
fit2call(fit)
}
