% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassify.R
\name{ggplotROC}
\alias{ggplotROC}
\title{Draw ROC curves}
\usage{
ggplotROC(y, predlist, labels, ...)
}
\arguments{
\item{y}{A numeric vector}

\item{predlist}{A list of numeric vecors as results of predict.glm}

\item{labels}{Character}

\item{...}{further arguments to be passed to geom_line()}
}
\description{
Draw ROC curves
}
\examples{
form1=paste0("AMD~",paste0(colnames(ExampleData)[3:10],collapse="+"))
form2=paste0("AMD~",paste0(colnames(ExampleData)[3:16],collapse="+"))
fit1=glm(as.formula(form1),data=ExampleData,family=binomial)
fit2=glm(as.formula(form2),data=ExampleData,family=binomial)
pred1=predict(fit1,type="response")
pred2=predict(fit2,type="response")
labels <- c("without genetic factors", "with genetic factors")
ggplotROC(ExampleData$AMD,predlist=list(pred1,pred2),labels=labels)
}
