% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassify.R
\name{plotROC2}
\alias{plotROC2}
\title{Draw ROC curves}
\usage{
plotROC2(data, cOutcome, predrisk, labels)
}
\arguments{
\item{data}{A data.frame}

\item{cOutcome}{Column number of the outcome variable.}

\item{predrisk}{Vector of predicted risk}

\item{labels}{Labels given to the ROC curves.}
}
\description{
Draw ROC curves
}
\examples{
form1=paste0("AMD~",paste0(colnames(ExampleData)[3:10],collapse="+"))
form2=paste0("AMD~",paste0(colnames(ExampleData)[3:16],collapse="+"))
fit1=glm(as.formula(form1),data=ExampleData,family=binomial)
fit2=glm(as.formula(form2),data=ExampleData,family=binomial)
pred1=predict(fit1,type="response")
pred2=predict(fit2,type="response")
labels <- c("without genetic factors", "with genetic factors")
plotROC2(data=ExampleData, cOutcome=2, predrisk=cbind(pred1,pred2), labels=labels)
}
