% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassify.R
\name{predictPlot}
\alias{predictPlot}
\title{Draw a Predictiveness Plot}
\usage{
predictPlot(fitlist, labels, size, ...)
}
\arguments{
\item{fitlist}{A list of objects of class glm}

\item{labels}{String Optional labels}

\item{size}{numeric line size}

\item{...}{Further arguments to be passed to geom_line()}
}
\value{
A ggplot
}
\description{
Draw a Predictiveness Plot
}
\examples{
form1=paste0("AMD~",paste0(colnames(ExampleData)[3:10],collapse="+"))
form2=paste0("AMD~",paste0(colnames(ExampleData)[3:16],collapse="+"))
fit1=glm(as.formula(form1),data=ExampleData,family=binomial)
fit2=glm(as.formula(form2),data=ExampleData,family=binomial)
labels=c("without genetic factors", "with genetic factors")
predictPlot(list(fit1,fit2),labels=labels)
}
